2023 day 5, CPU go brr, brain go durrr

+++ 2023/src/puzzles.h
@@ -59,8 +59,8 @@ public:
 
 class Day5 : public Puzzle {
 public:
-  int solve_part1(std::string input);
-  int solve_part2(std::string input);
+  long long solve_part1(std::string input);
+  long long solve_part2(std::string input);
   void solve(std::string input) override;
 };
